Steven Registrar y Login

+++ ProyectoA/src/proyectoa/bd/CrudArchivo.java
@@ -6,9 +6,11 @@ package proyectoa.bd;
 
 import guiA.datos.Usuario;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.util.HashMap;
 import java.util.logging.Level;
@@ -23,30 +25,64 @@ import java.util.logging.Logger;
 public class CrudArchivo {
     public static String rutaBase=System.getProperty("user.home")+File.separator+"bd";
 
-    public static void crearUsuario(Object datos, String nombreArchivo)throws Exception{
-        ObjectOutputStream guardarObjeto = null;
-        try {
-          File archivoRuta = new File(rutaBase);
-         if(!archivoRuta.exists()){
-          archivoRuta.mkdir();
-          }
-         File archivo = new File(rutaBase, nombreArchivo+".dat");
-           if(!archivo.exists()){
-             archivo.createNewFile();
-            }
-         FileOutputStream guardarArchivo=new FileOutputStream(archivo);
-         guardarObjeto=new ObjectOutputStream(guardarArchivo);
-         guardarObjeto.writeObject(datos);
-        } catch (FileNotFoundException ex) {
-            throw new Exception("El archivo no exite");
-        } catch (IOException ex) {
-            throw new Exception("Error al guardar los datos");
-        }finally{
-            try{
-                guardarObjeto.close();
-            }catch(IOException ex){
-                Logger.getLogger(CrudArchivo.class.getName()).log(Level.SEVERE, null, ex);
-            }
+ public static void crearUsuario(Object datos, String nombreArchivo) throws IOException {
+    ObjectOutputStream guardarObjeto = null;
+    try {
+        File archivoRuta = new File(rutaBase);
+        if (!archivoRuta.exists()) {
+            archivoRuta.mkdir();
+        }
+
+        File archivo = new File(rutaBase, nombreArchivo);
+        if (archivo.exists()) {
+            throw new IOException("El archivo ya existe");
+        }
+
+        FileOutputStream guardarArchivo = new FileOutputStream(archivo);
+        guardarObjeto = new ObjectOutputStream(guardarArchivo);
+        guardarObjeto.writeObject(datos);
+    } finally {
+        if (guardarObjeto != null) {
+            guardarObjeto.close();
         }
     }
+}
+ 
+
+
+public static Object buscarObjeto(String nombreArchivo) throws IOException, ClassNotFoundException {
+    ObjectInputStream leerObjeto = null;
+    Object objetoEncontrado = null;
+    try {
+        File archivo = new File(rutaBase, nombreArchivo + ".dat");
+        if (!archivo.exists()) {
+            throw new IOException("El archivo no existe");
+        }
+
+        FileInputStream leerArchivo = new FileInputStream(archivo);
+        leerObjeto = new ObjectInputStream(leerArchivo);
+        objetoEncontrado = leerObjeto.readObject();
+    } finally {
+        if (leerObjeto != null) {
+            leerObjeto.close();
+        }
+    }
+    return objetoEncontrado;
+}
+
+ public static void cargarUsuarios() {
+        try {
+            // Buscar los usuarios en el archivo y asignarlos al mapa de usuarios
+            Usuario.usuariosNuevos = (HashMap<String, Usuario>) buscarObjeto("Usuarios");
+        } catch (IOException | ClassNotFoundException e) {
+            // Manejar el error si no se pueden cargar los usuarios
+            System.err.println("Error al cargar usuarios: " + e.getMessage());
+            e.printStackTrace();
+        }
+    }
+
+
+   
+
+
 }
